[devel/ycm] Fix build stage trying to download files
On some systems, the build system is trying to download files from remote
repositories and fails because those files have now changed. This is something
that is not allowed in robotpkg anyway and was only partially addressed by
patch-aa. Disabling 3rdparty package fixes the remaining bits. This
package is only used by middlware/yarp, and this does not seem to create any
issue.

Bump PKGREVISION