[mk] Fix behaviour when multiple goals are specified on the command line
In commit:1b3fc02, the rule to setup the trusted environment was simplified
so that target failures do report the actual target name and not the
environment wrapper target name (for clarity). However, this introduced a
bug when multiple targets are specified on the command line (e.g. with
`make update confirm`): robotpkg would recurse as many times as the number of
specified goals (thus twice for `make update confirm`, which would update a
package twice in a row).

In order to keep the desired behaviour of printing the actual target name
that failed, the fix consists in partially reverting commit:1b3fc02 but
renaming the recursive target to the joined string of all command line
targets with a `-' separator. In the most common case of a single target,
this is a noop. For multiple targets, like 'update confirm', this will print
errors about the target 'update-confirm', which is a good trade-off.