[mk/fetch] Add a MASTER_SITES syntax to avoid DISTFILES concatenation
Sometimes, especially on github, is it convenient to specify a URL
in MASTER_SITES that directly downloads a DISTFILES
(e.g. http://github.com/<acct>/<project>/archive/<tag>.tar.gz) but still save
it as DISTNAME (e.g. <pkg>-<version>.tar.gz).
This saves some declarations in the Makefile to put the DISTFILES in a
subdirectory or similar.

This functionality is enabled for MASTER_SITES that start with a - (hyphen).