[interfaces/ros2-rosidl] Import ros2-rosidl-3.1.1
The rosidl API consists of a few message related static functions and types
along with a definition of what code should be generated by messages in
different languages. The generated message code specified in the API will be
language specific, but may or may not reuse generated code for other
languages. The generated message code specified in the API contains things like
the message data structure, functions for construction, destruction, etc. The
API will also implement a way to get the type support structure for the message
type, which is used when publishing or subscribing to a topic of that message
type.