[meta-pkg/ros2-core] Import ros_core-0.9.3
The Robot Operating System (ROS) is a set of software libraries and tools for
building robot applications. From drivers to state-of-the-art algorithms, and
with powerful developer tools, ROS has what you need for your next robotics
project. And it’s all open source.

Since ROS was started in 2007, a lot has changed in the robotics and ROS
community. The goal of the ROS 2 project is to adapt to these changes,
leveraging what is great about ROS 1 and improving what isn’t.