[devel/foonathan-memory] Import foonathan-memory-0.7.1
The C++ STL allocator model has various flaws. For example, they are fixed to a
certain type, because they are almost necessarily required to be templates. So
you can't easily share a single allocator for multiple types. In addition, you
can only get a copy from the containers and not the original allocator
object. At least with C++11 they are allowed to be stateful and so can be made
object not instance based. But still, the model has many flaws. Over the course
of the years many solutions have been proposed, for example EASTL. This library
is another. But instead of trying to change the STL, it works with the current
implementation.