[middleware/sros2] Import sros2-0.10.4
This package provides the tools and instructions to use ROS2 on top of
DDS-Security. The security feature is tested across platforms (Linux, macOS,
and Windows) as well as across different languages (C++ and Python).

Although SROS2 is designed to work with any secure middleware, at the moment it
is tested with RTI Connext Secure 5.3.1 and eProsima's Fast-RTPS 1.6.0.