[mk/sysdep] Teach robotpkg about gtk3 (and gdk-pixbuf2 and shared-mime-info)
Both gtk2 and gtk3 have a indirect dependency on shared-mime-info.pc via
gdk-pixbuf-2.0.pc (this might not be true for older gtk-2.0 < 2.36 or similar,
but well ...).

Since shared-mime-info installs its .pc file in shared/pkgconfig instead of
lib, the cleanest way to get PKG_CONFIG_PATH correct is to add the
indirect dependencies in mk/sysdep.