[sysutils/py-rosdistro] Fix SYSTEM_SEARCH on ubuntu2204
There is a mess with ros2 and python3-rosdistro on ubuntu2204.
Both a system python3-rosdistro and a ros2 python3-rosdistro-modules exist and
are conflicting. It happens that ros2 is built with the latter, that must thus
be installed, but lacks the various script normally provided by this package.

So to overcome all this, don't check for the scripts in SYSTEM_SEARCH...