[middleware/omniORB] Remove omniORB.cfg configuration file from PLIST
The configuration file installation is managed by a INSTALL script, so it
does not belong to PLIST. Especially since the DEINSTALL script leaves the
configuration file behind if it was modified.

In addition, the PLIST was hardcoding the path and did not account for a
modified PKG_SYSCONFDIR.

Reported by Joseph Mirabel, thanks!