[mk/sysdep] export SETUPTOOLS_USE_DISTUTILS=stdlib to fix setuptools mess
According to setuptools changelog, starting from setuptools 50.0.0, the python
distutils core is overriden by setuptools. This breaks most of the ROS PLISTs
(as they use distutils), due to flat vs. non-flat python eggs ... but the
behaviour actually depends on the distribution: there is no uniform
behaviour accross differents distributions and different setuptools
versions. For instance, debian does not seem affected and fedora is the most
affected.

So far, the simplest is to export SETUPTOOLS_USE_DISTUTILS=stdlib, which is
the documented obscure workaround and seems to actually calm things down.
Until next time?