[mk/internal] Add a 'sh' macro to fix 'shell' in log-filtered context
In a rule where the LOGFILTER is enabled, SHELL is redefined to output to the
logging script. This breaks 'shell' macros called in that context, since stdout
is redirected.

So add a 'sh' macro that restores a normal shell (${SH} -c) before invoking
'shell'. This takes advantage of the 'foreach' function that binds its
variable locally so that the target-specific SHELL can be overriden. The
'let' function would have been more suited, but it's not supported by older
GNU make still in use.