[devel/ros2-python-cmake-module] Set PYTHON_EXT_SUFFIX consistently
Make sure that the package variable PythonExtra_EXTENSION_SUFFIX is set
to robotpkg PYTHON_EXT_SUFFIX, according to the system in use. Update patch-aa
to reflect this.

When using a vanilla system pacakge, an empty PYTHON_EXT_SUFFIX will be
rejected (see patch-aa). But on systems where there is a python-cmake-module
system package, the definition is not supposed to be empty.