[mk/build] Update for gmake-4.4 compatibility
Due to rework of MAKEOVERRIDES handling in gmake-4.4, the behaviour is
changed: exporting an empty MAKEOVERRIDES in the environment prevents any
further recursive make process to inherit command lines variables.
Not clear if it is a feature or a bug, but anyway ...

Using target specific variable attached to the 'do-build' and 'do-install'
action fixes this and also works with earlier versions of gmake.