[pkgtools/pkg_install] Don't include pgp signature support
pkg_install ability to sign packages relies on netpgpverify, which is not
included in robotpkg yet. Since binary robotpkg packages are not normally
fetched online but rather built from source, signing binary packages is not on
the top requirement list.

Delay this feature for now, this can easily be added later by reverting this
commit.