[meta-pkgs/boost] Delete outdated boost package and make it sysdep-only
The boost package in robotpkg was very much obsolete and thus quite
useless. And there is nowadays no need for a boost package in robotpkg:
it is available as a recent-enough system package on all supported
platforms.

Also, with the widespread availability of C++11 and later versions, there
is less pressure on boost since the libstdc++ has integrated much of its
features.

Update dependant packages includes and while there sort a few unsorted
dependency lines.