[mk/sysdep] Teach robotpkg about matio
Matio is an open-source C library for reading and writing binary MATLAB MAT
files. This library is designed for use by programs/libraries that do not have
access or do not want to rely on MATLAB's shared libraries.

Note: matio is provided in robotpkg/wip/, but is nowadays available as a
system package everywhere. So it is imported in robotpkg as a sysdep only and
will be later removed in wip/.

OK gsaurel@