[devel/ros2-python-cmake-module] Always use empty PYTHON_SOABI
Since python-cmake-module>=0.12, PYTHON_SOABI is dropped and python libraries
are compiled as a bare .so.

Upgrading to 0.12 is currently breaking things (Python.h not found), probably
because some other package needs an upgrade too. So in the meantime, modify
patch-aa to always use an empty PYTHON_SOABI. This will make PLISTs simpler.