[interfaces/ros2-rosidl-python] Update to 0.21.1
Changes since 0.20:
* Install compiled libraries only to 'lib' (#195)

This version installs one file less, so require it as a minimum version to
avoid dealing with PLIST complexity. This corresponds to the packaged version
in ubuntu-22.04.

Also, deal with python-cmake-module<0.12 when used as a system package (only
ubuntu-22.04) that will compile python libraries with a PYTHON_EXT_SUFFIX
extension. In order to avoid too much PLIST complexity, create a PLIST filter
for this case, adding the required PYTHON_EXT_SUFFIX.