[mk/depend] Fix prefixsearch.sh when SYSTEM_PREFIX contains `@' characters
prefixsearch.sh was using sed substitutions with the `@' delimiter to perform
some replacements in strings containing paths components. Since `@' are valid
characters in filenames, this was failing for users having such characters in
their PATH.

Since this was used only to replace a `%' by the matched path in SYSTEM_SEARCH
specifications, use a simple dedicated shell function for doing the job.
This saves the burden of using sed again to quote any @ in the replacement
strings.

For emulating sed s///g, the shell function is recursive, but the number of
recursions is not expected to exceed 1 or 2 since there are no reason to
replace more than one % character in practice.