[mk/sysdep] Fix gcc>=7 version detection
Since gcc>=7, gcc -dumpversion may return only the major version.
A new -dumpfullversion was added, returning the ... full version.

It seems that when given either -dumpversion or -dumpfullversion, gcc ignores
any other options even if they are invalid. So using both may preserve
compatibility with gcc<7 (not that it is particularly needed, but well).